Daño final

+++ Models/Syncro.cs
@@ -9,7 +9,7 @@ namespace Syncro.Models
     public class Syncro
     {
         public long Nivel { get; set; }
-        public long DañoBase { get; set; }
-        public long DañoFinal { get; set; }
+        public double DañoBase { get; set; }
+        public double DañoFinal { get; set; }
     }
 }
